refactor (domain) : LottoRank 의 getRank의 matchingCount 를 int에서 long로 변경

+++ src/main/java/lotto/domain/LottoRank.java
@@ -32,7 +32,7 @@ public enum LottoRank {
         return matchingCount;
     }
 
-    public static LottoRank getRank(int matchingCount, boolean isSecond) {
+    public static LottoRank getRank(long matchingCount, boolean isSecond) {
         if ((!isSecond) && matchingCount == THIRD.getMatchingCount()) {
             return THIRD;
         }
